+++ desk/pages/main/index.tsx
@@ -1,7 +1,7 @@
-import Best from '@/src/components/units/main/best/Best.container'
-import Recent from '@/src/components/units/main/recent/Recent.container'
-import Youtube from '@/src/components/units/main/youtube/Youtube.container'
-import BoardsRegisterButton from '@/src/components/units/main/boardsRegisterButton'
+import Best from '@/src/components/units/main/categories/best/Best.container'
+import Recent from '@/src/components/units/main/categories/recent/Recent.container'
+import Youtube from '@/src/components/units/main/categories/youtube/Youtube.container'
+import BoardsRegisterButton from '@/src/components/units/main/components/boardsRegisterButton'
 
 export default function MainPage() {
   return (
