+++ desk/src/components/units/main/categories/best/Best.presenter.tsx
@@ -1,13 +1,13 @@
 import { Center } from '@chakra-ui/react'
 import MainImageStyle from '@/src/components/ui/mainImageStyle'
-import CategoryHeader from '../categoryHeader/CategoryHeader.container'
+import CategoryHeader from '../../components/categoryHeader/CategoryHeader.container'
 
-export default function LikeUI() {
-  const categoryTitle = 'Like'
+export default function BestUI() {
+  const categoryTitle = 'Best'
 
   return (
     <>
-      <CategoryHeader categoryTitle={categoryTitle} moreButtonHidden={true} />
+      <CategoryHeader categoryTitle={categoryTitle} moreButtonHidden={false} />
       <Center m={2}>
         {/* 이미지 UI 구성을 위한 임시 Key값. 추후 api 연결 시 수정 예정 */}
         {[1, 1, 1, 1].map(num => (
