+++ desk/src/components/units/main/categories/youtube/Youtube.container.tsx
@@ -1,9 +1,9 @@
 import { useQuery } from '@apollo/client'
-import { FETCH_YOUTUBE } from './Youtube.queries'
-import { TQuery, TYoutube } from '@/src/commons/types/generated/types'
-import YoutubeUI, { YoutubeUIProps } from './Youtube.presenter'
 import ErrorMessage from '@/src/components/ui/errorMessage'
 import CustomSpinner from '@/src/components/ui/spinner'
+import YoutubeUI, { YoutubeUIProps } from './Youtube.presenter'
+import { TQuery, TYoutube } from '@/src/commons/types/generated/types'
+import { FETCH_YOUTUBE } from './Youtube.queries'
 
 export default function Youtube() {
   const { loading, error, data } = useQuery<Pick<TQuery, 'fetchYoutube'>>(FETCH_YOUTUBE)
